D/encode moved from server to lib
